+++ gpu.go
@@ -1,10 +1,10 @@
 package main
 
 type GPUtype struct {
-	Model    string `json:"model"`
-	Used     bool   `json:"used"`
-	MemTotal int64  `json:"mem_total"`
-	Util     int64  `json:"util"`
-	Mem      int64  `json:"mem"`
-	Temp     int64  `json:"temp"`
+	Model    string  `json:"model"`
+	Used     bool    `json:"used"`
+	MemTotal int64   `json:"mem_total"`
+	Util     float64 `json:"util"`
+	Mem      float64 `json:"mem"`
+	Temp     float64 `json:"temp"`
 }
