+++ mem.go
@@ -1,7 +1,7 @@
 package main
 
 type MEMtype struct {
-	Total int64 `json:"total"`
-	Alloc int64 `json:"alloc"`
-	Util  int64 `json:"util"`
+	Total int64   `json:"total"`
+	Alloc int64   `json:"alloc"`
+	Util  float64 `json:"util"`
 }
